+++ packages/form_bloc/lib/form_bloc.dart
@@ -5,4 +5,5 @@ export 'src/blocs/field/field_bloc.dart';
 export 'src/blocs/text_field/text_field_bloc.dart';
 export 'src/blocs/select_field/select_field_bloc.dart';
 export 'src/blocs/boolean_field/boolean_field_bloc.dart';
+export 'src/blocs/file_field/file_field_bloc.dart';
 export 'src/validators.dart';