+++ packages/form_bloc/lib/src/blocs/file_field/file_field_state.dart
@@ -1,15 +1,16 @@
+import 'dart:io';
 import 'package:meta/meta.dart';
 import 'package:form_bloc/form_bloc.dart';
 
-class BooleanFieldBlocState extends FieldBlocState<bool> {
+class FileFieldBlocState extends FieldBlocState<File> {
   final String _toStringName;
   final bool isRequired;
 
   @override
-  bool get isValid => !isRequired ? true : value;
+  bool get isValid => isRequired ? value != null : true;
 
-  BooleanFieldBlocState({
-    @required bool value,
+  FileFieldBlocState({
+    @required File value,
     @required bool isInitial,
     @required String toStringName,
     @required this.isRequired,
@@ -20,9 +21,9 @@ class BooleanFieldBlocState extends FieldBlocState<bool> {
           additionalProps: <dynamic>[isRequired],
         );
 
-  BooleanFieldBlocState copyWith({bool value, bool isInitial}) {
-    return BooleanFieldBlocState(
-      value: value ?? this.value,
+  FileFieldBlocState withValue(File value, {bool isInitial}) {
+    return FileFieldBlocState(
+      value: value,
       isInitial: isInitial ?? this.isInitial,
       toStringName: _toStringName,
       isRequired: isRequired,
@@ -39,7 +40,7 @@ class BooleanFieldBlocState extends FieldBlocState<bool> {
     }
     _toString += ' {';
     if (isInitial) _toString += ' isInitial: $isInitial,';
-    _toString += ' value: $value';
+    _toString += ' value: ${value?.path}';
     _toString += ' }';
 
     return _toString;