+++ packages/form_bloc/lib/src/blocs/form/form_event.dart
@@ -1,4 +1,5 @@
 import 'package:equatable/equatable.dart';
+import 'form_state.dart';
 
 abstract class FormBlocEvent extends Equatable {
   FormBlocEvent([List props = const <dynamic>[]]) : super(props);
@@ -6,10 +7,17 @@ abstract class FormBlocEvent extends Equatable {
 
 class SubmitFormBloc extends FormBlocEvent {}
 
-class UpdateFormBloc extends FormBlocEvent {
+class UpdateFormBlocState<SuccessResponse, FailureResponse>
+    extends FormBlocEvent {
+  final FormBlocState<SuccessResponse, FailureResponse> state;
+
+  UpdateFormBlocState(this.state) : super(<dynamic>[state]);
+}
+
+class UpdateFormBlocStateIsValid extends FormBlocEvent {
   final bool isValid;
 
-  UpdateFormBloc(this.isValid) : super(<dynamic>[isValid]);
+  UpdateFormBlocStateIsValid(this.isValid) : super(<dynamic>[isValid]);
 }
 
 class LoadFormBloc extends FormBlocEvent {}
@@ -17,3 +25,5 @@ class LoadFormBloc extends FormBlocEvent {}
 class ReloadFormBloc extends FormBlocEvent {}
 
 class ClearFormBloc extends FormBlocEvent {}
+
+class CancelSubmissionFormBloc extends FormBlocEvent {}