+++ packages/form_bloc/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.3.0
+
+* Dependency and Documentation Updates.
+* Add `submissionProgress` property to `FormBlocState`.
+* Add `canSubmit` property to `FormBlocState`.
+* Add `FormBlocSubmissionFailed` state to `FormBloc`.
+* Add `FormBlocSubmissionCancelled` state to `FormBloc`.
+* Add `cancelSubmission` event to `FormBloc`.
+* Add `updateState` event to `FormBloc`.
+* Add `onCancelSubmission` method to `FormBloc`.
+* Add `FileFieldBloc`.
+
 # 0.2.0
 
 Initial Version of the library.